requirements.txt
blinker==1.8.2
click==8.1.7
flask==3.0.3
flask-mail==0.10.0
importlib-metadata==8.2.0
itsdangerous==2.2.0
jinja2==3.1.4
lxml==5.3.0
MarkupSafe==2.1.5
python-docx==1.1.2
tqdm==4.66.5
typing-extensions==4.12.2
werkzeug==3.0.3
zipp==3.19.2
